+++ modules/extensions-xml-support/src/main/java/org/mule/runtime/extension/internal/loader/xml/validator/GlobalElementNamesValidator.java
@@ -4,7 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.runtime.extension.internal.feeder.validator;
+package org.mule.runtime.extension.internal.loader.xml.validator;
 
 import static java.lang.String.format;
 import static org.mule.runtime.internal.util.NameValidationUtil.verifyStringDoesNotContainsReservedCharacters;
